requirements.txt
mss==9.0.1
numpy==1.26.0
opencv-python==4.8.0.76
Pillow==10.0.1
